refactor(release): move the publication guard to the publish primitive
The source-size growth ratchet charged 45 new lines: release.ts is 2000+ lines
and every line above 1500 counts, so three call sites plus a helper landed
entirely in the charged region. Raising the baseline is a defect, so the fix is
extraction.

guard_publication and create_and_push_annotated_tag now live in
release_publication.ts (720 lines, free), which already owns die and _push_tag.
The tag sequence is one function because it is one irreversible act with one
precondition — splitting it across call sites is what let the resumed-push path
reach _push_tag having read nothing. Behaviour unchanged; release.ts nets to
zero and the ratchet reads exactly at baseline.
