+++ src/scripts/release.ts
@@ -122,7 +122,6 @@ import {
     HEAD_LABELS,
     collect_span_commits,
     derive_category_hits,
-    publication_blockers,
     render_derived_head_values,
 } from './_lib/release_highlights.js';
 import {
@@ -191,6 +190,8 @@ import {
     _tag_exists_local,
     _tag_exists_remote,
     _target_from_branch,
+    create_and_push_annotated_tag,
+    guard_publication,
     die,
     gh,
     git,
@@ -1182,31 +1183,6 @@ function _step(n: number, total: number, msg: string): void {
     process.stdout.write(`[${n}/${total}] ${msg}\n`);
 }
 
-/**
- * Refuse an irreversible publication whose section is not publishable.
- *
- * Called at the THREE call sites that publish, never inside a formatter:
- * annotated-tag creation, the resumed push of a tag created but never pushed,
- * and the GitHub Release body. A pure formatter has no notion of whether it is
- * actually publishing, which is why an earlier attempt at that placement was
- * refused; the call site does, and it is the one that can still stop.
- *
- * `die` throws before the command runs, so nothing irreversible fires after a
- * refusal — that ordering is what `tests/scripts/release_publication_guard.test.ts`
- * asserts against the drill's recorded command list rather than by reading this
- * comment.
- */
-function _refuse_unpublishable(sectionBody: string, version: string, surface: string): void {
-    const blockers = publication_blockers(sectionBody, version);
-    if (blockers.length === 0) {
-        return;
-    }
-    die(
-        `refusing to publish the ${surface} for ${version} — ` +
-            `${String(blockers.length)} publication blocker(s):\n` +
-            blockers.map((b) => `    - ${b}`).join('\n'),
-    );
-}
 
 function execute(
     plan: Plan,
@@ -1417,19 +1393,7 @@ function execute(
             _step(8, total, `Tag ${plan.target} already on ${REMOTE} — skip`);
         } else {
             _step(8, total, `Tag ${plan.target} exists locally — push only`);
-            // 3.2 — the recorded bypass. The changelog is read only in the
-            // CREATION branch below, so a resume over a tag that was created
-            // but never pushed reached `_push_tag` having read nothing. A
-            // guard that covers creation alone misses exactly this path, and
-            // pushing the tag is as irreversible as creating it.
-            const staged = extract_changelog_section(read_changelog_text(), plan.target);
-            if (!staged) {
-                die(
-                    `CHANGELOG.md on ${MAIN_BRANCH} carries no section for ${plan.target} — ` +
-                        'refusing to push a tag whose changelog entry is missing',
-                );
-            }
-            _refuse_unpublishable(staged!.body, plan.target, 'tag push (resumed)');
+            guard_publication(plan.target, 'tag push (resumed)');
             _push_tag(plan.target);
         }
     } else {
@@ -1440,16 +1404,7 @@ function execute(
         // tag replaces the previous lightweight one so tag metadata is a
         // fourth surface carrying the same single-source content.
         _step(8, total, `Tag merge commit (annotated, from merged CHANGELOG) and push ${plan.target}`);
-        const merged = extract_changelog_section(read_changelog_text(), plan.target);
-        if (!merged) {
-            die(
-                `CHANGELOG.md on ${MAIN_BRANCH} carries no section for ${plan.target} — ` +
-                    'refusing to tag a release whose changelog entry is missing',
-            );
-        }
-        _refuse_unpublishable(merged!.body, plan.target, 'annotated tag');
-        run(['git', 'tag', '-a', plan.target, '-m', tag_message_from_section(merged!.body, plan.target)]);
-        _push_tag(plan.target);
+        create_and_push_annotated_tag(plan.target);
     }
 
     // ─── 9. GitHub Release ──────────────────────────────────────────────────
@@ -1479,7 +1434,7 @@ function execute(
             GH_RELEASE_NOTES_LIMIT,
             '`CHANGELOG.md`',
         );
-        _refuse_unpublishable(tagged_section!.body, plan.target, 'GitHub Release notes');
+        guard_publication(plan.target, 'GitHub Release notes');
         gh(['release', 'create', plan.target, '--title', plan.target, '--notes', notes]);
 
         // ─── 9b. dispatch-chain the tag-triggered workflows (--ci only) ──────
